refactor settingsscreen
